requirements/base.txt
SQLAlchemy
alembic
Grab
pytesseract
Pillow
apscheduler